+++ fastly/config.go
@@ -3,10 +3,13 @@ package fastly
 import (
 	"fmt"
 
+	gofastly "github.com/fastly/go-fastly/fastly"
 	"github.com/hashicorp/terraform/helper/logging"
-	"github.com/hashicorp/terraform/terraform"
-	gofastly "github.com/sethvargo/go-fastly/fastly"
+	"github.com/hashicorp/terraform/httpclient"
+	"github.com/terraform-providers/terraform-provider-fastly/version"
 )
+
+const TerraformProviderProductUserAgent = "terraform-provider-fastly"
 
 type Config struct {
 	ApiKey  string
@@ -24,7 +27,9 @@ func (c *Config) Client() (interface{}, error) {
 		return nil, fmt.Errorf("[Err] No API key for Fastly")
 	}
 
-	gofastly.UserAgent = terraform.UserAgentString()
+	composedUserAgent := fmt.Sprintf("%s %s/%s", httpclient.UserAgentString(), TerraformProviderProductUserAgent, version.ProviderVersion)
+
+	gofastly.UserAgent = composedUserAgent
 
 	fastlyClient, err := gofastly.NewClientForEndpoint(c.ApiKey, c.BaseURL)
 	if err != nil {